+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/slider
 
+## 1.0.2
+
+### Patch Changes
+
+- Add styles props
+
 ## 1.0.1
 
 ### Patch Changes